% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_greengenesDb.R
\name{getGreenGenes13.5Db}
\alias{getGreenGenes13.5Db}
\title{getGreenGenes13.5Db download and build database}
\usage{
getGreenGenes13.5Db(pkgname = "greengenes13.5MgDb")
}
\description{
getGreenGenes13.5Db download and build database
}
\examples{
getGreenGenes13.5Db()
}

