% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg13.5MgDb.R
\name{gg13.5MgDb}
\alias{gg13.5MgDb}
\title{MgDb-class object for greengenes version 13.5}
\description{
gg13.5MgDb is a MgDb-class object with
sequence and taxonomic data for the greengenes
database version 13.5, \url{http://greengenes.secondgenome.com}
}
\examples{
gg13.5MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
