% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengenes13.5MgDb.R
\docType{package}
\name{greengenes13.5MgDb-package}
\alias{greengenes13.5MgDb-package}
\title{Greengenes 16S rRNA database MgDb Package.}
\description{
This package loads a \code{\link[metagenomeFeatures]{MgDb-class}} object with
sequence and taxonomic data for the greengenes
database version 13.5, \url{http://greengenes.secondgenome.com}.
Vignettes for working with \code{\link[metagenomeFeatures]{MgDb-class}}
objects are in the \pkg{metagenomeFeatures} package.
}
\examples{
ls('package:greengenes13.5MgDb')
gg13.5MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\keyword{package}
